requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.2
django-analytical==3.1.0
django-crispy-forms==2.3
django-multiselectfield==0.1.13
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
tzdata==2024.2
